requirements.txt
openai-agents
gradio
fastapi
uvicorn
python-docx
yfinance
matplotlib
google-generativeai
huggingface_hub 